+++ CHANGELOG.md
@@ -13,6 +13,9 @@
   * 新增 `ct`（`Contain`）运算符，用于取代 `in`（`Include`）运算符（使用 `in` 将输出警告）
 * Bean Searcher Boot Starter
   * 默认自动配置添加 `EnumFieldConvertor`
+  * 新增 `bean-searcher.field-convertor.use-enum` 配置项，表示是否自动添加 `EnumFieldConvertor`，默认 `true`
+  * 新增 `bean-searcher.use-map-searcher` 配置项，表示是否自动创建 `MapSearcher` 检索器，默认 `true`
+  * 新增 `bean-searcher.use-bean-searcher` 配置项，表示是否自动创建 `BeanSearcher` 检索器，默认 `true`
 
 # v3.1.2 & v3.0.4 @ 2021-12-02